backend/TTS/requirements.txt
numpy
torch
torchaudio
pydantic
fastapi
requests
modal
chatterbox-tts